#11UpdateDataPoskoAndHaveRelationWithTrc

+++ app/Http/Controllers/BencanaController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Bencana;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use RealRashid\SweetAlert\Facades\Alert;
 
@@ -16,8 +16,8 @@ class BencanaController extends Controller
      */
     public function index()
     {
-        $bencana = Bencana::select(DB::raw("concat(tanggal,' ',waktu) as waktu"),'tanggal as tgl','waktu as time','bencana.id as idBencana','nama','lokasi', 'status', 'posko_id as posko', 'pengungsi_id as pengungsi', 'updated_at as waktuUpdate')
-            // ->leftJoin('posko AS p', 'bencana.posko_id', '=', 'p.id')
+        $bencana = Bencana::select(DB::raw("concat(tanggal,' ',waktu) as waktu"), 'tanggal as tgl', 'waktu as time', 'bencana.id as idBencana', 'nama', 'lokasi', 'status', 'posko_id as posko', 'pengungsi_id as pengungsi', 'updated_at as waktuUpdate')
+        // ->leftJoin('posko AS p', 'bencana.posko_id', '=', 'p.id')
             ->orderBy('bencana.tanggal', 'desc')
             ->paginate(5);
         return view('admin.bencana.index', [
@@ -33,7 +33,7 @@ class BencanaController extends Controller
      */
     public function createBencana(Request $request)
     {
-        if (auth()->user()->hasAnyRole(['pusdalop', 'trc'])) {
+        if (auth()->user()->hasAnyRole(['pusdalop'])) {
             // $request->validate([
             //     'namaDepan' => ['required', 'max:50'],
             //     'namaBelakang' => ['required', 'max:50'],
@@ -86,7 +86,7 @@ class BencanaController extends Controller
     {
         $bencana = Bencana::where('id', $id)->first();
 
-        if (auth()->user()->hasAnyRole(['pusdalop', 'trc'])) {
+        if (auth()->user()->hasAnyRole(['pusdalop'])) {
             $bencana->nama = $request->namaBencana;
             $bencana->tanggal = $request->tanggal;
             $bencana->waktu = $request->waktu;
@@ -111,22 +111,26 @@ class BencanaController extends Controller
         //
     }
 
-    public function delete($id){
-        $delete = Bencana::destroy($id);
+    public function delete($id)
+    {
+        if (auth()->user()->hasAnyRole(['pusdalop'])) {
+            $delete = Bencana::destroy($id);
 
-        // check data deleted or not
-        if ($delete == 1) {
-            $success = true;
-            $message = "Data berhasil dihapus";
-        } else {
-            $success = true;
-            $message = "Data gagal dihapus";
+            // check data deleted or not
+            if ($delete == 1) {
+                $success = true;
+                $message = "Data berhasil dihapus";
+            } else {
+                $success = true;
+                $message = "Data gagal dihapus";
+            }
+
+            //  return response
+            return response()->json([
+                'success' => $success,
+                'message' => $message,
+            ]);
         }
-
-        //  return response
-        return response()->json([
-            'success' => $success,
-            'message' => $message,
-        ]);
+        return back();
     }
 }
